update fix upload avatar

+++ src/middlewares/createTrainning.middleware.ts
@@ -1,13 +1,13 @@
-import e, { Request, Response, NextFunction } from 'express';
+import { Request, Response, NextFunction } from 'express';
 import cloudinary from 'cloudinary';
 import dotenv from 'dotenv';
 import fs from 'fs';
 dotenv.config();
 const ENV = process.env;
 cloudinary.v2.config({
-    cloud_name: ENV.CLOUD_NAME,
-    api_key: ENV.API_KEY_CLOUD,
-    api_secret: ENV.API_SECRET_CLOUD,
+    cloud_name: ENV.CLOUD_DINARY_NAME,
+    api_key: ENV.API_KEY_CLOUD_DINARY,
+    api_secret: ENV.API_SECRET_CLOUD_DINARY,
 });
 const create_media_path = async (
     req: Request,
@@ -15,34 +15,24 @@ const create_media_path = async (
     next: NextFunction,
 ) => {
     try {
-        const files = req.files as Express.Multer.File[];
-        const urlMedias: any[] = [];
-        const uploadPromises = files.map(async (file: any) => {
-            const result = await cloudinary.v2.uploader.upload(file.path, {
-                resource_type: 'auto',
-            });
-            fs.unlink(file.path, () => {
-                console.log('deleted path');
-            });
-            return result;
+        const file = req.file as Express.Multer.File;
+        if (!file) {
+            return res
+                .status(400)
+                .json({ success: false, message: 'No file uploaded' });
+        }
+        const result = await cloudinary.v2.uploader.upload(file.path, {
+            resource_type: 'auto',
         });
-        Promise.all(uploadPromises)
-            .then(results => {
-                const urls = results.map(result => result.secure_url);
-                urlMedias.push(urls);
-                req.body.media_path = urlMedias;
-                next();
-            })
-            .catch(error => {
-                res.status(203).json({
-                    success: false,
-                    message: error?.message,
-                });
-            });
+        fs.unlink(file.path, () => {
+            console.log('deleted path');
+        });
+        req.body.avatar = result.secure_url;
+        next();
     } catch (error: any) {
         res.status(500).json({
             success: false,
-            message: 'server error: ' + error.mesage,
+            message: 'server error: ' + error.message,
         });
     }
 };
